BrDlVtxGen 0x10022600, BrDlVtxLit 0x10021C70: match the original's float arithmetic
The whole-image run found a 1-ulp colour difference at frame 345. Symbolic
x87 evaluation of both binaries then showed two association differences a
byte-score search had introduced:
- BrDlVtxLit's first light row added (m1*dy + m2*dz) + m0*dx; the original
  adds (m1*dy + m0*dx) + m2*dz.
- BrDlVtxGen's transform rows added ((x + z) + y) + w; the original adds
  ((y + z) + x) + w. VC5 orders these adds by operand kind, not source
  order; (double)pSrc->x gives the original's order.
Every float both functions store is now the original's expression tree,
rounding points included. Ledgers re-run at the new numbers.

+++ src/core/drawing/br_dlvtx_texgen.c
@@ -105,7 +105,7 @@ extern float       DAT_105cd9fc;
 extern float       DAT_105ccd48;
 extern float DAT_105cd9fc;
 /* Transcription notes (2026-09-24, by hand from the asm; compiled /O2 /Op
- * like the rest of its original TU): 50 of 362 instructions still differ
+ * like the rest of its original TU): 52 of 362 instructions still differ
  * (difflib count).  Spellings that each moved a whole class (all measured):
  *  - the lights are N64 Light records read as BYTES: VC5 merges col[0]/col[1]
  *    into one dword load + `and`/`mov dl,ah`, exactly as the original;
@@ -132,23 +132,26 @@ extern float DAT_105cd9fc;
  *    declares the TU's other light/clip globals: the symbol table in front
  *    of the function decides x87 operand-role ties.  Setup order, groupings
  *    and declaration order were found by a hill-climb over those choices.
- * Open (x87 scheduling / operand role only; every row pairs):
- *  - the transform rows: the original issues the x product LAST (column
- *    loads z, y, then the vertex x); ours issues it second.  Dead: all six
- *    term orders, row/association grouping (8 forms), x as deref / cast /
- *    indexed, x column absolute / array / extern position (65 positions);
- *  - m[0]*n0: the original loads m[0]; ours loads n0.  Dead: struct matrix
- *    pointer, const pointer, `*m`;
- *  - the light setup's three extra fxch (direction fild issued before the
- *    last colour store): dead under all 720 statement orders;
- *  - the texgen spill slots for dotX and the t-scale cast swap (0x2c/0x30).
+ *  - `(double)pSrc->x` in the transform rows (0x100221D0's spelling): it
+ *    is what makes VC5 add the three products in the original's order,
+ *    ((y + z) + x) + w.  Source order does not: VC5 reorders these adds by
+ *    operand kind, and the plain spelling adds ((x + z) + y) + w -- different
+ *    rounding, a real divergence the byte score hid (found 2026-09-24).
+ * Arithmetic check: every float the function stores, with its rounding
+ * points, is the same expression tree as the original's (symbolic x87
+ * evaluation of both binaries).
+ * Open (x87 scheduling and register choice only): fxch/load order in the
+ * transform rows, the m[0]*n0 operand role, the light setup's three extra
+ * fxch, the texgen spill slots (0x2c/0x30).
  * @t4-pass 0x10022600 1 2026-09-24 probes 733 bytes 1206 insns 359 regions 10 rows 11 census no  (light setup: all 720 orders of the three direction and three colour statements; transform rows: 8 grouping/operand-kind forms of the x term; matrix/vertex pointer types: struct matrix pointer, const, `*m`, cast and indexed x; nothing moved)
- * @t4-pass 0x10022600 2 2026-09-24 probes 272 bytes 1206 insns 359 regions 10 rows 11 census yes  (census: every row is an x87 operand-role or issue-slot pair plus the dotX/t-scale spill-slot swap, which is what declaration order decides -- moved each of the 17 locals to every other slot; nothing moved) */
+ * @t4-pass 0x10022600 2 2026-09-24 probes 272 bytes 1206 insns 359 regions 10 rows 11 census yes  (census: every row is an x87 operand-role or issue-slot pair plus the dotX/t-scale spill-slot swap, which is what declaration order decides -- moved each of the 17 locals to every other slot; nothing moved)
+ * @t4-pass 0x10022600 3 2026-09-24 probes 272 bytes 1214 insns 363 regions 10 rows 29 census yes  (after the arithmetic fix; census: the rows are x87 load order and register choice, which declaration order decides -- moved each local to every other slot; nothing moved)
+ * @t4-pass 0x10022600 4 2026-09-24 probes 720 bytes 1214 insns 363 regions 10 rows 29 census no  (all 720 orders of the light setup's direction and colour statements; nothing moved) */
 /* @t3 0x10022600 2026-09-24 -- CERTIFIED COMPLETE, NOT BYTE-EXACT.
- * @t3-measure bytes 1206/1212 insns 359/362 rows 7+4 regions 10 oracle EQUIVALENT
- * @t3-effort passes 2 zero-movement 1 2
- * Residue is x87 scheduling and operand role only (see the Open list above);
- * every row pairs.  Do not reopen before the end-grind (CLAUDE.md rule 12).
+ * @t3-measure bytes 1214/1212 insns 363/362 rows 14+15 regions 10 oracle EQUIVALENT
+ * @t3-effort passes 4 zero-movement 3 4
+ * Residue is x87 scheduling and register choice only (see the Open list
+ * above); the arithmetic is the original's, rounding points included.  Do not reopen before the end-grind (CLAUDE.md rule 12).
  * Oracle coverage: the 25 scripts run 253 of the 362 instructions -- never
  * the light refresh (the cache is always valid on entry here) -- and the
  * game's one light has equal x/y/z direction bytes.  So the refresh was also
@@ -210,10 +213,10 @@ const uint8_t *BrDlVtxGen(const uint8_t *p)
     pVc = pV;
     for (i = 0; i < n; i++) {
         pn = &pSrc->n1;
-        pV[i].cx = DAT_105d1780 * pn[-4] + DAT_105d1770 * pn[-5] + pSrc->x * DAT_105d1760 + DAT_105d1790;
-        pV[i].cy = DAT_105d1784 * pn[-4] + DAT_105d1774 * pn[-5] + pSrc->x * DAT_105d1764 + DAT_105d1794;
-        pV[i].cz = DAT_105d1788 * pn[-4] + DAT_105d1778 * pn[-5] + pSrc->x * DAT_105d1768 + DAT_105d1798;
-        pV[i].cw = DAT_105d178c * pn[-4] + DAT_105d177c * pn[-5] + pSrc->x * DAT_105d176c + DAT_105d179c;
+        pV[i].cx = DAT_105d1780 * pn[-4] + DAT_105d1770 * pn[-5] + (double)pSrc->x * DAT_105d1760 + DAT_105d1790;
+        pV[i].cy = DAT_105d1784 * pn[-4] + DAT_105d1774 * pn[-5] + (double)pSrc->x * DAT_105d1764 + DAT_105d1794;
+        pV[i].cz = DAT_105d1788 * pn[-4] + DAT_105d1778 * pn[-5] + (double)pSrc->x * DAT_105d1768 + DAT_105d1798;
+        pV[i].cw = DAT_105d178c * pn[-4] + DAT_105d177c * pn[-5] + (double)pSrc->x * DAT_105d176c + DAT_105d179c;
 
         if (DAT_100a9a50 != 0) {
             off = (DAT_100a9a50 - 1) << 6;
